Move route execution logic to _internal/kori folder
• Move route-executor.ts and route-registry.ts to _internal/kori/
• Update import paths in kori-internal.ts and fetch-handler-factory.ts
• Remove empty route-executor directory
• Consolidate internal implementation in single location
