--- packages/kori/src/_internal/route-executor/index.ts
@@ -1,2 +1,0 @@
-export { createRouteExecutor } from './route-executor.js';
-export { createRouteRegistry, type RouteRecord } from './route-registry.js';
